making units more consistent

+++ h/Units.hpp
@@ -5,7 +5,7 @@ constexpr double pixelsByMeter = 25;
 double pixelsToMeters(int pixels);
 double pixelsToMeters(double pixels);
 
-double metersToPixels(double meters);
+int metersToPixels(double meters);
 
 Vector2D pixelsToMeters(Vector2D pixels);
 Vector2D metersToPixels(Vector2D meters);
@@ -20,9 +20,9 @@ inline double pixelsToMeters(double pixels)
    return pixels / pixelsByMeter;
 }
 
-inline double metersToPixels(double meters)
+inline int metersToPixels(double meters)
 {
-   return meters * pixelsByMeter;
+   return static_cast<int>(meters * pixelsByMeter);
 }
 
 inline Vector2D pixelsToMeters(Vector2D pixels)
